+++ packages/project-access/src/project/search.ts
@@ -198,27 +198,41 @@ async function findRootsForPath(path: string): Promise<{ appRoot: string; projec
             };
         }
         // Project must be CAP, find project root
-        try {
-            const { root } = parse(appRoot);
-            let projectRoot = dirname(appRoot);
-            while (projectRoot !== root) {
-                if (await getCapProjectType(projectRoot)) {
-                    // We have found a CAP project as root. Check if the found app is not directly in CAP's 'app/' folder.
-                    // Sometime there is a <CAP_ROOT>/app/package.json file that is used for app router (not an app)
-                    if (join(projectRoot, 'app') !== appRoot) {
-                        return {
-                            appRoot,
-                            projectRoot
-                        };
-                    }
-                }
-                projectRoot = dirname(projectRoot);
-            }
-        } catch {
-            // No project root can be found at parent folder.
+        const projectRoot = await findCapProjectRoot(appRoot);
+        if (projectRoot) {
+            return {
+                appRoot,
+                projectRoot
+            };
         }
     } catch {
         // Finding root should not throw error. Return null instead.
+    }
+    return null;
+}
+
+/**
+ * Find CAP project root path.
+ *
+ * @param path - path inside CAP project
+ * @returns - CAP project root path
+ */
+async function findCapProjectRoot(path: string): Promise<string | null> {
+    try {
+        const { root } = parse(path);
+        let projectRoot = dirname(path);
+        while (projectRoot !== root) {
+            if (await getCapProjectType(projectRoot)) {
+                // We have found a CAP project as root. Check if the found app is not directly in CAP's 'app/' folder.
+                // Sometime there is a <CAP_ROOT>/app/package.json file that is used for app router (not an app)
+                if (join(projectRoot, 'app') !== path) {
+                    return projectRoot;
+                }
+            }
+            projectRoot = dirname(projectRoot);
+        }
+    } catch {
+        // No project root can be found at parent folder.
     }
     return null;
 }
@@ -420,8 +434,8 @@ export async function findCapProjects(options: {
     readonly wsFolders: WorkspaceFolder[] | string[];
 }): Promise<string[]> {
     const result = new Set<string>();
-    const excludeFolders = ['node_modules', 'dist', 'src', 'webapp', 'MDKModule', 'gen'];
-    const fileNames = [FileName.Pom, FileName.Package];
+    const excludeFolders = ['node_modules', 'dist', 'webapp', 'MDKModule', 'gen'];
+    const fileNames = [FileName.Pom, FileName.Package, FileName.CapJavaApplicationYaml];
     const wsRoots = wsFoldersToRootPaths(options.wsFolders);
     for (const root of wsRoots) {
         const filesToCheck = await findBy({
@@ -429,7 +443,26 @@ export async function findCapProjects(options: {
             root,
             excludeFolders
         });
-        const foldersToCheck = Array.from(new Set(filesToCheck.map((file) => dirname(file)))); //only directories without duplicates
+        const appYamlsToCheck = Array.from(
+            new Set(
+                filesToCheck
+                    .filter((file) => basename(file) === FileName.CapJavaApplicationYaml)
+                    .map((file) => dirname(file))
+            )
+        );
+        const foldersToCheck = Array.from(
+            new Set(
+                filesToCheck
+                    .filter((file) => basename(file) !== FileName.CapJavaApplicationYaml)
+                    .map((file) => dirname(file))
+            )
+        );
+        for (const appYamlToCheck of appYamlsToCheck) {
+            const capRoot = await findCapProjectRoot(appYamlToCheck);
+            if (capRoot) {
+                result.add(capRoot);
+            }
+        }
         for (const folderToCheck of foldersToCheck) {
             if ((await getCapProjectType(folderToCheck)) !== undefined) {
                 result.add(folderToCheck);