+++ packages/xml-odata-annotation-converter/test/parser/escaping.test.ts
@@ -1,0 +1,8 @@
+import { removeEscapeSequences } from '../../src/parser/escaping';
+
+describe('removeEscapeSequences', () => {
+    it('', () => {
+        const result = removeEscapeSequences('--->&apos;&gt;&lt;&amp;&quot;<---');
+        expect(result).toMatchInlineSnapshot(`"--->'><&\\"<---"`);
+    });
+});